Add import to __init__.py
