+++ pyvista/examples/__init__.py
@@ -4,3 +4,4 @@ from pyvista.examples.downloads import *
 from pyvista.examples.examples import *
 from . import gltf
 from . import hdr
+from . import vrml
